backend/requirements.txt
# AetherNexus Backend - Python Dependencies
# Установка: pip install -r requirements.txt

# ============================================
# Core Framework
# ============================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# ============================================
# API & Web
# ============================================
python-multipart==0.0.6
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1

# ============================================
# Message Broker
# ============================================
# Apache Kafka
kafka-python==2.0.2
aiokafka==0.10.0

# Redis (Pub/Sub & Cache)
redis==5.0.1
hiredis==2.2.3
aioredis==2.0.1

# ============================================
# Vector Database - Qdrant
# ============================================
qdrant-client==1.7.0

# ============================================
# Graph Database - Neo4j
# ============================================
neo4j==5.15.0

# ============================================
# ML & Embeddings
# ============================================
# Для генерации эмбеддингов
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1
numpy==1.26.2

# Альтернатива: OpenAI embeddings (опционально)
# openai==1.3.7

# ============================================
# Text Processing & NLP
# ============================================
nltk==3.8.1
spacy==3.7.2
# Для русского языка: python -m spacy download ru_core_news_sm
# Для английского: python -m spacy download en_core_web_sm

# ============================================
# Code Parsing & Analysis
# ============================================
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-java==0.20.2
tree-sitter-javascript==0.20.1
tree-sitter-typescript==0.20.3

# AST parsing
astunparse==1.6.3

# ============================================
# Git Integration
# ============================================
GitPython==3.1.40
pygit2==1.12.2

# ============================================
# External Integrations
# ============================================
# Jira
jira==3.5.2
atlassian-python-api==3.41.5

# Confluence
atlassian-python-api==3.41.5

# Slack
slack-sdk==3.23.0
slack-bolt==1.18.0

# ============================================
# Database & ORM
# ============================================
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9  # PostgreSQL (опционально)

# ============================================
# Search & Indexing
# ============================================
whoosh==2.7.4  # Full-text search
elasticsearch==8.11.0  # Альтернатива (опционально)

# ============================================
# Utilities
# ============================================
python-dotenv==1.0.0
pyyaml==6.0.1
toml==0.10.2
python-json-logger==2.0.7
structlog==23.2.0

# ============================================
# Testing
# ============================================
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2  # Для тестирования API
faker==20.1.0

# ============================================
# Development Tools
# ============================================
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0
pylint==3.0.2

# ============================================
# Monitoring & Logging
# ============================================
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# ============================================
# Security
# ============================================
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1

# ============================================
# Async & Concurrency
# ============================================
# asyncio - встроенный модуль Python, не требует установки
aiofiles==23.2.1
celery==5.3.4  # Для фоновых задач (опционально)

# ============================================
# Data Processing
# ============================================
pandas==2.1.3
python-dateutil==2.8.2

# ============================================
# Version Pinning (для стабильности)
# ============================================
# Python версия: 3.10+
# Установка: python3.10 -m pip install -r requirements.txt

